use new style for hlint pragmas
Given that the library itself uses such identifiers, it seemed in
keeping with its style to leave the variable names as-is and just use
the new style for the pragmas. It also seems plausible that just
disabling the warning with a compiler flag could be preferred, so a
branch doing that is also provided.